% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_messages.R
\name{activity_time_summary}
\alias{activity_time_summary}
\title{summary of activities over time}
\usage{
activity_time_summary(action_times, dataI = NULL, dataF = NULL)
}
\arguments{
\item{dataI}{minimum date to be considered (POSIXct format)}

\item{dataF}{maximum date to be considered (POSIXct format)}

\item{action_time}{a vector of POSIXct elements refering to the time of the activities.}
}
\description{
funtions which decrive the activities over time
}
