% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixPath.R
\name{fixPath}
\alias{fixPath}
\title{fixPath}
\usage{
fixPath(file.path, silent = FALSE)
}
\arguments{
\item{file.path}{path}

\item{silent}{FALSE}
}
\value{
correct file.path
}
\description{
Validity check (and fix) for files path
}
\author{
Livio Finos
}

