% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_messages.R
\name{get_group_threads}
\alias{get_group_threads}
\title{get infos from threads}
\usage{
get_group_threads(mess)
}
\arguments{
\item{mess}{a data.frame as given by \code{getMessages}}
}
\description{
\code{get_owner_from_threads} return the name of the user, \code{get_group_threads} return a TRUE/FALSE factor indicating if the chat coms from a group (TRUE) or a personal (FALSE) chat.
}
\author{
Caterina Ciampanelli, Anna Petrina, Livio Finos
}

