% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnagrafica.R
\name{getAnagrafica}
\alias{getAnagrafica}
\title{conoscere le informazioni del profilo}
\usage{
getAnagrafica(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
dataset (1x20) contenente le informazioni del profilo
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna l'insieme dei dati dell'utente in un'unica riga di dataset
in particolare restituisce: sesso, n. di sessioni e di accessi, n. di amici di varia categoria, n. di post,
n. di attività del wall delle varie categorie, n. di eventi con varia risposta, n. di foto, n. di posti e n. di messaggi privati.
}
\author{
Davide Meneghetti
}

