% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnagraficaData.R
\name{getAnagraficaData}
\alias{getAnagraficaData}
\title{conoscere le informazioni del profilo}
\usage{
getAnagraficaData(percorso, data)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{data}{data di riferimento}
}
\value{
dataset (1x20) contenente le informazioni del profilo dopo la data di rifermento
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna l'insieme dei dati di un profilo in un'unica riga di dataset
dopo una certa data
in particolare restituisce: sesso, n. di sessioni e di accessi, n. di amici di varia categoria, n. di post,
n. di attività del wall delle varie categorie, n. di eventi con varia risposta, n. di foto, n. di posti e n. di messaggi privati.
}
\author{
Davide Meneghetti
}

