% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataReg.R
\name{getDataReg}
\alias{getDataReg}
\title{data di registrazione del profilo}
\usage{
getDataReg(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
data: data di registrazione del profilo
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna la data di registrazione
}
\author{
Davide Meneghetti
}
