% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDati.R
\name{getDati}
\alias{getDati}
\title{Dati di un profilo facebook}
\usage{
getDati(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
diverse informazioni
}
\description{
Funzione che dato il percorso dei dati del profilo facebook restituisce una serie di informazioni sul profilo
}
\author{
Davide Meneghetti
}

