% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventi.R
\name{getEventi}
\alias{getEventi}
\title{conoscere il numero di eventi}
\usage{
getEventi(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
dataset (1x4) contenente il numero dei vari tipi di eventi
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di eventi di vario tipo
tipi di evento: confermato, forse, rifiutato, non risposto.
Questa fuzione è parzialmente funzionante anche con i profili in lingua inglese e spagnola
}
\author{
Davide Meneghetti
}

