% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventiData.R
\name{getEventiData}
\alias{getEventiData}
\title{conoscere il numero di eventi}
\usage{
getEventiData(percorso, dataRef)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataRef}{data di riferimento}
}
\value{
dataset (1x4) contenente il numero dei vari tipi di eventi dopo la data di riferimento
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di eventi di vario tipo dopo una certa data
tipi di evento: confermato, forse, rifiutato, non risposto.
}
\author{
Davide Meneghetti
}

