% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEvents.R
\name{getEvents}
\alias{getEvents}
\title{conoscere il numero di eventi}
\usage{
getEvents(percorso, dataI, dataF)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataI}{data di inizio di riferimento}

\item{dataF}{data di fine di riferimento}
}
\value{
dataset (1x4) contenente il numero dei vari tipi di eventi dopo la data di riferimento
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di eventi di vario tipo compresi fra 2 date
tipi di evento: confermato, forse, rifiutato, non risposto.
}
\author{
Davide Meneghetti
}
