% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFriends.R
\name{getFriends}
\alias{getFriends}
\title{amici di un profilo facebook}
\usage{
getFriends(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
dataset interi(1x4): numero di amici accettati,
                             numero di richieste di amicizia effettuate ma non risposte,
                             numero di richieste di amicizia ricevute ma non risposte,
                             numero di amici rimossi
}
\description{
Funzione che dato il percorso dei dati del profilo facebook il numero di amici(per tipologia)
}
\author{
Davide Meneghetti
}
