% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfo.R
\name{getInfo}
\alias{getInfo}
\title{indicazione del sesso del profilo}
\usage{
getInfo(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
F,M
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna alcune infomrazioni personali (dalla pagina settings.html)
}
\author{
Davide Meneghetti, Livio Finos
}
