% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMessages_summary.R
\name{getMessages_summary}
\alias{getMessages_summary}
\title{getMessages_summary}
\usage{
getMessages_summary(mess)
}
\arguments{
\item{mess}{result of a call of \code{getMessages}.}

\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
a list with few summary statistics
}
\author{
Caterina Ciampanelli, Livio Finos
}
