% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMessagesTexts.R
\name{getMessagesTexts}
\alias{getMessagesTexts}
\title{daset dei messaggi privati}
\usage{
getMessagesTexts(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
dataset dei messaggi composto da data, utente autore e testo dei messaggi
}
\description{
DA RIVEDERE
Funzione che dato il percorso dei dati del profilo facebook ritorna il dataset dei messaggi privati inviati e ricevuti
comprendendo 3 colonne: la data, l'utente e il testo del messaggio
}
\author{
Davide Meneghetti, Livio Finos
}

