% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMessaggi.R
\name{getMessaggi}
\alias{getMessaggi}
\title{conoscere il numero di messaggi privati}
\usage{
getMessaggi(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
numero di messaggi privati
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di messaggi privati
}
\author{
Davide Meneghetti
}

