% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMessaggiData.R
\name{getMessaggiData}
\alias{getMessaggiData}
\title{conoscere il numero di messaggi privati dopo una data}
\usage{
getMessaggiData(percorso, dataRef)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataRef}{data di riferimento}
}
\value{
numero di messaggi privati successi alla data di riferimento indicata
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di messaggi privati dopo una data di riferimento
}
\author{
Davide Meneghetti
}

