% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMessaggiData2.R
\name{getMessaggiData2}
\alias{getMessaggiData2}
\title{conoscere il numero di messaggi privati dopo una data}
\usage{
getMessaggiData2(percorso, dataI, dataF)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataI}{data di inizio di riferimento}

\item{dataF}{data di fine di riferimento}
}
\value{
numero di messaggi privati compresi fra 2 date
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di messaggi privati compresi fra 2 date
}
\author{
Davide Meneghetti
}

