% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNAccessi.R
\name{getNAccessi}
\alias{getNAccessi}
\title{conoscere il numero di sessioni aperte e il numero accessi}
\usage{
getNAccessi(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
dataset(1x2): numero di sessioni e numero di accessi
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di sessioni aperte e il numero accessi
}
\author{
Davide Meneghetti
}

