% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNAccessiData.R
\name{getNAccessiData}
\alias{getNAccessiData}
\title{conoscere il numero sessioni aperte e di accessi}
\usage{
getNAccessiData(percorso, dataRef)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataRef}{data di riferimento}
}
\value{
dataset (1x2) contenente:numero di sessioni aperte e di accessi dopo la data di riferimento
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero sessioni aperte e di accessi dopo una certa data
}
\author{
Davide Meneghetti
}

