% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFoto.R
\name{getNFoto}
\alias{getNFoto}
\title{conoscere il numero di foto}
\usage{
getNFoto(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
numero di foto pubblicate dal profilo
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di foto
}
\author{
Davide Meneghetti
}

