% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNFotoData.R
\name{getNFotoData}
\alias{getNFotoData}
\title{conoscere il numero di foto}
\usage{
getNFotoData(percorso, data)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{data}{data di riferimento}
}
\value{
numero di foto pubblicate dal profilo dopo la data di riferimento
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di foto dopo una certa data
}
\author{
Davide Meneghetti
}

