% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNFotoData2.R
\name{getNFotoData2}
\alias{getNFotoData2}
\title{conoscere il numero di foto}
\usage{
getNFotoData2(percorso, dataI, dataF)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataI}{data di inizio di riferimento}

\item{dataF}{data di fine di riferimento}
}
\value{
numero di foto pubblicate dal profilo dopo la data di riferimento comprese fra 2 date
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di foto comprese fra 2 date
}
\author{
Davide Meneghetti
}

