% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNPost.R
\name{getNPost}
\alias{getNPost}
\title{conoscere il numero di post}
\usage{
getNPost(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
numero di post scritti sulla bacheca
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di post sulla bacheca
}
\author{
Davide Meneghetti
}

