% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNPostData.R
\name{getNPostData}
\alias{getNPostData}
\title{conoscere il numero di post}
\usage{
getNPostData(percorso, dataRef)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataRef}{data di riferimento}
}
\value{
numero di post scritti sulla bacheca dopo la data di riferimento
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di post sulla bacheca dopo una certa data
}
\author{
Davide Meneghetti
}

