% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNPostData2.R
\name{getNPostData2}
\alias{getNPostData2}
\title{conoscere il numero di post}
\usage{
getNPostData2(percorso, dataI, dataF)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataI}{data di inizio di riferimento}

\item{dataF}{data di fine di riferimento}
}
\value{
numero di post scritti sulla bacheca compresi fra 2 date
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di post sulla bacheca compresi fra 2 date
}
\author{
Davide Meneghetti
}

