% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlaces.R
\name{getPlaces}
\alias{getPlaces}
\title{conoscere il numero di posti}
\usage{
getPlaces(percorso, dataI, dataF)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataI}{data di inizio di riferimento}

\item{dataF}{data di fine di riferimento}
}
\value{
numero di posti dove ci si ? segnalati dopo la data di riferimento compresi fra 2 date
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di posti in cui si ? stati compresi fra 2 date
}
\author{
Davide Meneghetti, Livio Finos
}
