% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPosti.R
\name{getPosti}
\alias{getPosti}
\title{conoscere il numero di posti}
\usage{
getPosti(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
numero di posti dove ci si è segnalati
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di posti in cui si è stati
}
\author{
Davide Meneghetti
}

