% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPostiData.R
\name{getPostiData}
\alias{getPostiData}
\title{conoscere il numero di posti}
\usage{
getPostiData(percorso, dataRef)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataRef}{data di riferimento}
}
\value{
numero di posti dove ci si è segnalati dopo la data di riferimento
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di posti in cui si è stati dopo una certa data
}
\author{
Davide Meneghetti
}

