% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSesso.R
\name{getSesso}
\alias{getSesso}
\title{indicazione del sesso del profilo}
\usage{
getSesso(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
numero: 0 se femmina, 1 se maschio
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna l'indicazione del sesso
##'
}
\author{
Davide Meneghetti
}

