% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSex.R
\name{getSex}
\alias{getSex}
\title{indicazione del sesso del profilo}
\usage{
getSex(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
F,M
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna l'indicazione del sesso
##'
}
\author{
Davide Meneghetti
}

