% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWall.R
\name{getWall}
\alias{getWall}
\title{conoscere il numero delle varie attività sul wall}
\usage{
getWall(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
dataset (1x7) contenente le informazioni delle attività sul Wall
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero delle differenti attività che compaiono sul wall
tra cui amicizie, stati, like, piace, condivisioni, link e giocato.
Funzione valida anche per profili scritti in lingua inglese o spagnola.
}
\author{
Davide Meneghetti
}

