% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWallData.R
\name{getWallData}
\alias{getWallData}
\title{conoscere il numero delle varie attività sul wall}
\usage{
getWallData(percorso, dataRif)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataRif}{data di riferimento}
}
\value{
dataset (1x7) contenente le informazioni delle attività sul Wall dopo la data di riferimento
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero delle differenti attività che compaiono sul wall
tra cui amicizie, stati, like, piace, condivisioni, link e giocato dopo una certa data.
Funzione valida anche per profili scritti in lingua inglese o spagnola.
}
\author{
Davide Meneghetti
}

