% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWallPosts.R
\name{getWallPosts}
\alias{getWallPosts}
\title{Post scritti sul wall}
\usage{
getWallPosts(percorso)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}
}
\value{
vettore di testi
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna i post pubblicati sulla bacheca
}
\author{
Davide Meneghetti, Livio Finos
}

