% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini.R
\name{Gini}
\alias{Gini}
\title{Gini}
\usage{
Gini(user, n_mess = NULL, k = NULL, adjusted = TRUE)
}
\arguments{
\item{user}{}

\item{n_mess}{}

\item{k}{? il numero di modalit?, in questo caso il numero di amici totali che scrivono}
}
\description{
It computes Gini Index
}

