% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inDataIT.R
\name{inDataIT}
\alias{inDataIT}
\title{tradurre data da stringa per esteso a formato data}
\usage{
inDataIT(data)
}
\arguments{
\item{data}{stringa data estesa letta da facebook}
}
\value{
data in formato Date
}
\description{
Funzione traduce una data estesa letta come stringa da facebook in italiano e la restituisce in formato Date
}
\author{
Davide Meneghetti
}

