% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFBr.R
\docType{package}
\name{myFBr}
\alias{myFBr}
\alias{myFBr-package}
\title{myFBr: it Reads and visualizes the infos of your FaceBook profile}
\description{
The myFBr package provides three categories of important functions:
foo, bar and baz.
}
\section{myFBr functions}{

The foo functions ...
}

\examples{
path=system.file("extdata","example.zip",package = "myFBr")
getInfo(path)
}
