% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_over_all_folders.R
\name{perform_over_all_folders}
\alias{perform_over_all_folders}
\title{perform_over_all_folders}
\usage{
perform_over_all_folders(dir_list, function_to_run = function(cart)
  getMessages_summary(cart, dataI = dataI, dataF = dataF, stringhe = stringhe),
  dir_repos = "./temp_anna/", overwrite = FALSE)
}
\examples{
function_to_run <- function(cart) getMessages_summary(cart,dataI=dataI,dataF=dataF)
}
\author{
livio finos
}

