% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnagrafica.R, R/summary_texts.R
\name{getAnagrafica}
\alias{getAnagrafica}
\title{conoscere le informazioni del profilo}
\usage{
getAnagrafica(percorso, dataI = -Inf, dataF = +Inf)

getAnagrafica(percorso, dataI = -Inf, dataF = +Inf)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataI}{data di inizio di riferimento}

\item{dataF}{data di fine di riferimento}

\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataI}{data di inizio di riferimento}

\item{dataF}{data di fine di riferimento}
}
\value{
dataset (1x20) contenente le informazioni del profilo compresi fra 2 date

dataset (1x20) contenente le informazioni del profilo compresi fra 2 date
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna l'insieme dei dati di un profilo in un'unica riga di dataset
compresi fra 2 date in particolare restituisce: sesso, n. di sessioni e di accessi, n. di amici di varia categoria, n. di post,
n. di attivit? del wall delle varie categorie, n. di eventi con varia risposta, n. di foto, n. di posti e n. di messaggi privati.

Funzione che dato il percorso dei dati del profilo facebook ritorna l'insieme dei dati di un profilo in un'unica riga di dataset
compresi fra 2 date in particolare restituisce: sesso, n. di sessioni e di accessi, n. di amici di varia categoria, n. di post,
n. di attivit? del wall delle varie categorie, n. di eventi con varia risposta, n. di foto, n. di posti e n. di messaggi privati.
}
\author{
Davide Meneghetti

Davide Meneghetti
}

