% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMessages.R
\name{getMessages}
\alias{getMessages}
\alias{getNMessages}
\title{estre (il numero d)i messaggi privati}
\usage{
getMessages(percorso, dataI = NULL, dataF = NULL)
}
\arguments{
\item{percorso}{stringa che indica il percorso della cartella dei dati del profilo}

\item{dataI}{data di inizio di riferimento}

\item{dataF}{data di fine di riferimento}
}
\value{
numero di messaggi privati compresi fra 2 date
}
\description{
Funzione che dato il percorso dei dati del profilo facebook ritorna il numero di messaggi privati compresi fra 2 date
}
\author{
Davide Meneghetti, Livio Finos
}

