% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inDataIT.R
\name{inDataIT}
\alias{inDataIT}
\title{inDataIT}
\usage{
inDataIT(a)
}
\arguments{
\item{a}{stringa data estesa letta da facebook}
}
\value{
data in formato Date
}
\description{
traduce una data estesa letta come stringa da facebook in italiano e la restituisce in formato Date
}
\author{
Davide Meneghetti
}
